refactor: 컴포넌트 및 패키지 네이밍 변경
